third_party/blink/web_tests/fast/css/invalid-predefined-color-expected.txt

This test checks hexadecimal color parsing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(div, null).color is red
PASS getComputedStyle(div, null).color is red
PASS getComputedStyle(div, null).color is red
PASS getComputedStyle(div, null).color is red
PASS getComputedStyle(div, null).color is red
PASS getComputedStyle(div, null).color is black
PASS getComputedStyle(div, null).color is black
PASS getComputedStyle(div, null).color is black
PASS getComputedStyle(div, null).color is black
PASS getComputedStyle(div, null).color is black
PASS successfullyParsed is true

TEST COMPLETE

This style is "color: red"
This style is "color: FF0000"
This style is "color: #FF0000"
This style is "color: F00"
This style is "color: #F00"
This style is "color: #red"
This style is "color: #yellow"
This style is "color: #green"
This style is "color: #FFAALL"
This style is "color: #\#"
